Descomponetizacion de ImgLoader e integración en solo componente ReportForm
